plot out new design
I think it will make more sense to parse dialogue file into stream of events that user has control over. Which means that they will need to take care of printing and all that but they can control flow using iterator similar concept here. Implementing actions is probably what leads design this way - to make user pass a function somehow or to return user an instruction on when action takes place and make them call it when that happens